FINAL CLEANUP: Remove all CI/CD duplicates and consolidate
- Copy enhanced github-issues-workflow.js with create/update commands
- Remove entire ci-cd/ directory (all files were duplicates)
- Remove .modified empty file
- Repository now has single authoritative versions in .github/workflows/ and scripts/

🧹 CI/CD cleanup complete - no more duplicates
