--- ci-cd/deployment/firebase-cleanup.yml
@@ -1,47 +1,0 @@
-name: Cleanup Preview Channels
-
-on:
-  pull_request:
-    types: [closed]
-
-jobs:
-  cleanup:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      
-      - name: Setup Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: '22'
-          cache: 'npm'
-      
-      - name: Install Firebase CLI
-        run: npm install -g firebase-tools
-      
-      - name: Authenticate to Firebase
-        env:
-          FIREBASE_SERVICE_ACCOUNT: ${{ secrets.FIREBASE_SERVICE_ACCOUNT_BARTERVERSE_L9UQ3 }}
-        run: |
-          echo "$FIREBASE_SERVICE_ACCOUNT" > firebase-service-account.json
-          export GOOGLE_APPLICATION_CREDENTIALS=firebase-service-account.json
-      
-      - name: Delete Preview Channel
-        env:
-          GOOGLE_APPLICATION_CREDENTIALS: firebase-service-account.json
-        run: |
-          # Generate channel ID from PR number (same logic as preview action)
-          CHANNEL_ID="pr${{ github.event.pull_request.number }}"
-          
-          # Skip deletion if this is the testing channel (keep it for 7 days)
-          if [ "$CHANNEL_ID" = "testing" ] || [ "${{ github.head_ref }}" = "testing" ]; then
-            echo "Skipping deletion of testing channel - keeping for 7 days"
-            exit 0
-          fi
-          
-          # Delete the preview channel
-          firebase hosting:channel:delete $CHANNEL_ID --project barterverse-l9uq3 --site barterverse-auto-deploy --force || echo "Channel $CHANNEL_ID not found or already deleted"
-      
-      - name: Cleanup credentials
-        if: always()
-        run: rm -f firebase-service-account.json
